feat: refucturing code
