+++ src/assets/index.css
@@ -1,2 +1,2 @@
 @import url("./hero/index.css");
-@import url("./section/index.css");
+/* @import url("./section/index.css"); */
