--- src/assets/section/index.css
@@ -1,25 +1,0 @@
-.section_about {
-  background-color: rgb(0, 0, 0);
-
-  .wrapper__blob {
-    position: absolute;
-    inset: 0;
-    filter: blur(100px);
-    z-index: 0;
-
-    .blob {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      width: 65%;
-      height: 65%;
-      border-radius: 80%;
-      background: conic-gradient(
-        from 0deg,
-        rgba(0, 132, 255, 0.49),
-        rgba(255, 0, 0, 0.5)
-      );
-    }
-  }
-}
