+++ src/components/Home.jsx
@@ -1,28 +1,17 @@
-import { useState, useRef, useEffect } from "react";
+import React from "react";
 
-import HeroSection from "./Fragments/HeroSection";
-import Navbar from "./Fragments/Navbar";
-import About from "./Fragments/About";
-import Contact from "./Fragments/Contact";
+// Layouts
+import MainLayout from "./layouts/MainLayout";
+
+// Organisms
+import Sections from "./organisms/Sections";
 
 function Home() {
-  const aboutRef = useRef(null);
-  const heroRef = useRef(null);
-
-  const scrollToAbout = () => {
-    aboutRef.current?.scrollIntoView({ behavior: "smooth" });
-  };
-
-  const scrollToHero = () => {
-    heroRef.current?.scrollIntoView({ behavior: "smooth" });
-  };
-
   return (
     <>
-      <Navbar scrollToAbout={scrollToAbout} scrollToHero={scrollToHero} />
-      <HeroSection heroRef={heroRef} />
-      <About aboutRef={aboutRef} />
-      <Contact />
+      <MainLayout>
+        <Sections />
+      </MainLayout>
     </>
   );
 }
